+++ src/main/java/insane96mcp/progressivebosses/ai/wither/WitherRangedAttackGoal.java
@@ -18,17 +18,14 @@ public class WitherRangedAttackGoal extends Goal {
 	private final int attackInterval;
 	private final float attackRadius;
 	private final float attackRadiusSqr;
-	//Attack Interval multiplied by this when Wither's health drops below half
-	private final double attackSpeedMultiplier;
 	//Increases the rate of attack of the middle head the closer the player is to the wither
 	private final boolean increaseASOnNear;
 
-	public WitherRangedAttackGoal(WitherEntity wither, int attackInterval, float attackRadius, double attackSpeedMultiplier, boolean increaseASOnNear) {
+	public WitherRangedAttackGoal(WitherEntity wither, int attackInterval, float attackRadius, boolean increaseASOnNear) {
 		this.wither = wither;
 		this.attackInterval = attackInterval;
 		this.attackRadius = attackRadius;
 		this.attackRadiusSqr = attackRadius * attackRadius;
-		this.attackSpeedMultiplier = attackSpeedMultiplier;
 		this.increaseASOnNear = increaseASOnNear;
 		this.setMutexFlags(EnumSet.of(Goal.Flag.MOVE, Goal.Flag.LOOK));
 	}
@@ -103,7 +100,7 @@ public class WitherRangedAttackGoal extends Goal {
 				return;
 			witherTags.putInt(Strings.Tags.BARRAGE_ATTACK, barrageAttackTick - 1);
 			if (barrageAttackTick % 2 == 0) {
-				this.wither.launchWitherSkullToCoords(0, this.target.getPosX() + RandomHelper.getDouble(this.wither.getRNG(), -2d, 2d), this.target.getPosY() + (double)this.target.getEyeHeight() * 0.5D + RandomHelper.getDouble(this.wither.getRNG(), -2d, 2d), this.target.getPosZ() + RandomHelper.getDouble(this.wither.getRNG(), -2d, 2d), false);
+				this.wither.launchWitherSkullToCoords(RandomHelper.getInt(this.wither.getRNG(), 0, 3), this.target.getPosX() + RandomHelper.getDouble(this.wither.getRNG(), -2d, 2d), this.target.getPosY() + (double)this.target.getEyeHeight() * 0.5D + RandomHelper.getDouble(this.wither.getRNG(), -2d, 2d), this.target.getPosZ() + RandomHelper.getDouble(this.wither.getRNG(), -2d, 2d), false);
 			}
 		}
 		else if (--this.attackTime <= 0) {
@@ -111,19 +108,16 @@ public class WitherRangedAttackGoal extends Goal {
 				return;
 			if (RandomHelper.getFloat(this.wither.getRNG(), 0f, 1f) < .1f)
 				for (int h = 0; h < 3; h++) {
-					this.wither.launchWitherSkullToCoords(h, this.target.getPosX() + RandomHelper.getDouble(this.wither.getRNG(), -1.5d, 1.5d), this.target.getPosY() + RandomHelper.getDouble(this.wither.getRNG(), -1.5d, 1.5d) + (double)this.target.getEyeHeight() * 0.5D, target.getPosZ() + RandomHelper.getDouble(this.wither.getRNG(), -1.5d, 1.5d), RandomHelper.getDouble(this.wither.getRNG(), 0d, 1d) < 0.001F);
+					this.wither.launchWitherSkullToCoords(h, this.target.getPosX() + RandomHelper.getDouble(this.wither.getRNG(), -1.25d, 1.25d), this.target.getPosY() + RandomHelper.getDouble(this.wither.getRNG(), -1.25d, 1.25d) + (double)this.target.getEyeHeight() * 0.5D, target.getPosZ() + RandomHelper.getDouble(this.wither.getRNG(), -1.25d, 1.25d), RandomHelper.getDouble(this.wither.getRNG(), 0d, 1d) < 0.001F);
 				}
 			else
 				this.wither.launchWitherSkullToCoords(0, this.target.getPosX(), this.target.getPosY() + (double)this.target.getEyeHeight() * 0.5D, target.getPosZ(), RandomHelper.getDouble(this.wither.getRNG(), 0d, 1d) < 0.001F);
 			this.attackTime = this.attackInterval;
 
-			if (this.wither.isCharged() && this.attackSpeedMultiplier > 0d)
-				this.attackTime *= this.attackSpeedMultiplier;
-
 			if (this.increaseASOnNear) {
 				float distance = this.wither.getDistance(this.target);
 				if (distance < this.attackRadius) {
-					int nearBonusAS = (int) Math.round((this.attackInterval * 0.3d) * (1d - (distance / this.attackRadius)));
+					int nearBonusAS = (int) Math.round((this.attackInterval * 0.75d) * (1d - (distance / this.attackRadius)));
 					this.attackTime -= nearBonusAS;
 				}
 			}